feat: 소수점 연산 기능 추가

+++ practice/ex01/Lexer.hpp
@@ -25,7 +25,7 @@ class Lexer {
   Token parseOperatorToken(void);
 
   bool isOperator(char c);
-  int parseNumber(void);
+  double parseNumber(void);
 
   Lexer(Lexer const& lexer);
   Lexer& operator=(Lexer const& lexer);
