+++ practice/ex01/Parser.cpp
@@ -6,11 +6,11 @@ Parser::Parser(void) : lexer(), token(INVALID) {}
 Parser::~Parser(void) {}
 
 // Method
-int Parser::getResult(std::string input) {
+double Parser::getResult(std::string input) {
   lexer.setLexer(input);
   token = lexer.nextToken();
 
-  int value = expression();
+  double value = expression();
 
   if (token.type != END) {
     throw std::invalid_argument("Invalid token encountered!");
@@ -19,8 +19,8 @@ int Parser::getResult(std::string input) {
   return value;
 }
 
-int Parser::expression(void) {
-  int value = term();
+double Parser::expression(void) {
+  double value = term();
 
   while (token.type == PLUS || token.type == MINUS) {
     if (token.type == PLUS) {
@@ -35,16 +35,16 @@ int Parser::expression(void) {
   return value;
 }
 
-int Parser::factor(void) {
+double Parser::factor(void) {
   if (token.type == NUMBER) {
-    int value = token.value;
+    double value = token.value;
     token = lexer.nextToken();
     return value;
   }
 
   if (token.type == LPAREN) {
     token = lexer.nextToken();
-    int value = expression();
+    double value = expression();
     if (token.type != RPAREN) {
       throw std::invalid_argument("Invalid token encountered!");
     }
@@ -55,8 +55,8 @@ int Parser::factor(void) {
   throw std::invalid_argument("Invalid token encountered!");
 }
 
-int Parser::term(void) {
-  int value = factor();
+double Parser::term(void) {
+  double value = factor();
 
   while (token.type == MULTIPLY || token.type == DIVIDE) {
     if (token.type == MULTIPLY) {
@@ -64,7 +64,7 @@ int Parser::term(void) {
       value *= factor();
     } else {
       token = lexer.nextToken();
-      int f = factor();
+      double f = factor();
       if (f == 0) {
         throw std::invalid_argument("Devide by zero!");
       }
